@@ -12,20 +12,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.0.0",
-    "@angular/animations": "^17.0.0",
+    "@angular-devkit/build-angular": "^18.0.6",
+    "@angular/animations": "^18.0.5",
     "@angular/cdk": "^17.0.0",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
+    "@angular/common": "^18.0.5",
+    "@angular/compiler": "^18.0.5",
+    "@angular/core": "^18.0.5",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^17.0.0",
-    "@angular/localize": "^17.0.2",
+    "@angular/forms": "^18.0.5",
+    "@angular/localize": "^18.0.5",
     "@angular/material": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
-    "@angular/service-worker": "^17.0.2",
+    "@angular/platform-browser": "^18.0.5",
+    "@angular/platform-browser-dynamic": "^18.0.5",
+    "@angular/router": "^18.0.5",
+    "@angular/service-worker": "^18.0.5",
     "@apollo/client": "^3.0.0",
     "@auth0/angular-jwt": "^5.2.0",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
@@ -63,9 +63,9 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular/cli": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.2",
-    "@angular/language-service": "^17.0.2",
+    "@angular/cli": "^18.0.6",
+    "@angular/compiler-cli": "^18.0.5",
+    "@angular/language-service": "^18.0.5",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "coveralls": "^3.0.9",
@@ -79,6 +79,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~5.0.1",
     "tslint": "~6.1.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.4.5"
   }
 }
