@@ -9,7 +9,7 @@ import {ApiFiltersService} from './services/api-filters.service';
 import {AggregationService} from './services/aggregation.service';
 import { FilterComponent } from './shared/filter/filter.component';
 // import {NgxPaginationModule} from 'ngx-pagination'; <--- not used
-import {HttpClientModule} from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ActiveFilterComponent } from './shared/active-filter/active-filter.component';
 import { HomeComponent } from './home/home.component';
 import { OrganismComponent } from './organism/organism.component';
@@ -134,130 +134,122 @@ export function removeToken(): void {
 }
 
 // @ts-ignore
-@NgModule({
-  declarations: [
-    AppComponent,
-    FileComponent,
-    FilterComponent,
-    ActiveFilterComponent,
-    HomeComponent,
-    OrganismComponent,
-    SpecimenComponent,
-    DatasetComponent,
-    AnalysisComponent,
-    ArticleComponent,
-    HeaderComponent,
-    FileDetailComponent,
-    RobustLinkComponent,
-    RelatedItemsComponent,
-    OrganismDetailComponent,
-    SpecimenDetailComponent,
-    DatasetDetailComponent,
-    AnalysisDetailComponent,
-    ArticleDetailComponent,
-    SubprojectDetailComponent,
-    ProtocolSampleComponent,
-    ProtocolExperimentComponent,
-    ProtocolSampleDetailsComponent,
-    ProtocolExperimentDetailsComponent,
-    OrganismsSummaryComponent,
-    SpecimensSummaryComponent,
-    DatasetsSummaryComponent,
-    FilesSummaryComponent,
-    ProtocolAnalysisComponent,
-    NonExistingComponent,
-    RulesetSampleComponent,
-    RulesetExperimentComponent,
-    RulesetAnalysisComponent,
-    IndeterminateDirective,
-    FooterComponent,
-    ValidationSamplesComponent,
-    ValidationExperimentsComponent,
-    ValidationAnalysesComponent,
-    UsdaBovineComponent,
-    SheepatlasComponent,
-    SubprojectComponent,
-    TableClientSideComponent,
-    LoginComponent,
-    FilesUploadComponent,
-    OntologyImproverComponent,
-    OntologyDetailComponent,
-    ApiDocsComponent,
-    QueryLanguageComponent,
-    TrackhubsSubmissionComponent,
-    BulkFilesUploaderComponent,
-    EurofaangInfoComponent,
-    TableServerSideComponent,
-    NextflowSubmissionComponent,
-    ProtocolAnalysisDetailsComponent,
-    EnsemblAnnotationComponent,
-    LocalGenomeBrowserComponent,
-    GraphqlComponent,
-    GlobalSearchComponent,
-    DisplayDataComponent,
-    IndexFiltersComponent,
-    ShortenTitlePipe,
-    SubscriptionDialogComponent,
-    PortalDataTableComponent,
-
-    // to delete
-    CustomTableBuilderComponent,
-    // OntologyImproverWorkshopComponent,
-    // OntologyDetailWorkshopComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    // NgxPaginationModule,
-    NgxSpinnerModule,
-    // NgbDropdownModule,
-    BrowserAnimationsModule,
-    // ChartsModule,
-    NgChartsModule,
-    FormsModule,
-    FileUploadModule,
-    MatTableModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatTooltipModule,
-    MatButtonModule,
-    MatIconModule,
-    MatExpansionModule,
-    MatInputModule,
-    MatRadioModule,
-    MatDialogModule,
-    MatTabsModule,
-    MatSnackBarModule,
-    MatProgressSpinnerModule,
-    MatSelectModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatMenuModule,
-    MatCardModule,
-    FlexLayoutModule,
-    MatProgressBarModule,
-    MatToolbarModule,
-    MatBadgeModule,
-    MatTreeModule,
-    MatDividerModule,
-    MatCheckboxModule,
-    JwtModule.forRoot({
-      config: {
-        tokenGetter: getToken,
-        allowedDomains: ['api.aai.ebi.ac.uk'],
-        disallowedRoutes: ['https://api.aai.ebi.ac.uk/auth']
-      }
-    }),
-    // BsDropdownModule.forRoot(),
-    ServiceWorkerModule.register('ngsw-worker.js', {enabled: environment.production}),
-    NgxSmartModalModule.forRoot(),
-    // NgbModule,
-    NgcCookieConsentModule.forRoot(cookieConfig),
-    GraphQLModule,
-    MatListModule
-  ],
-  providers: [ApiDataService, AggregationService, SlicePipe, UserService, ApiFiltersService],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        FileComponent,
+        FilterComponent,
+        ActiveFilterComponent,
+        HomeComponent,
+        OrganismComponent,
+        SpecimenComponent,
+        DatasetComponent,
+        AnalysisComponent,
+        ArticleComponent,
+        HeaderComponent,
+        FileDetailComponent,
+        RobustLinkComponent,
+        RelatedItemsComponent,
+        OrganismDetailComponent,
+        SpecimenDetailComponent,
+        DatasetDetailComponent,
+        AnalysisDetailComponent,
+        ArticleDetailComponent,
+        SubprojectDetailComponent,
+        ProtocolSampleComponent,
+        ProtocolExperimentComponent,
+        ProtocolSampleDetailsComponent,
+        ProtocolExperimentDetailsComponent,
+        OrganismsSummaryComponent,
+        SpecimensSummaryComponent,
+        DatasetsSummaryComponent,
+        FilesSummaryComponent,
+        ProtocolAnalysisComponent,
+        NonExistingComponent,
+        RulesetSampleComponent,
+        RulesetExperimentComponent,
+        RulesetAnalysisComponent,
+        IndeterminateDirective,
+        FooterComponent,
+        ValidationSamplesComponent,
+        ValidationExperimentsComponent,
+        ValidationAnalysesComponent,
+        UsdaBovineComponent,
+        SheepatlasComponent,
+        SubprojectComponent,
+        TableClientSideComponent,
+        LoginComponent,
+        FilesUploadComponent,
+        OntologyImproverComponent,
+        OntologyDetailComponent,
+        ApiDocsComponent,
+        QueryLanguageComponent,
+        TrackhubsSubmissionComponent,
+        BulkFilesUploaderComponent,
+        EurofaangInfoComponent,
+        TableServerSideComponent,
+        NextflowSubmissionComponent,
+        ProtocolAnalysisDetailsComponent,
+        EnsemblAnnotationComponent,
+        LocalGenomeBrowserComponent,
+        GraphqlComponent,
+        GlobalSearchComponent,
+        DisplayDataComponent,
+        IndexFiltersComponent,
+        ShortenTitlePipe,
+        SubscriptionDialogComponent,
+        PortalDataTableComponent,
+        // to delete
+        CustomTableBuilderComponent,
+        // OntologyImproverWorkshopComponent,
+        // OntologyDetailWorkshopComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        // NgxPaginationModule,
+        NgxSpinnerModule,
+        // NgbDropdownModule,
+        BrowserAnimationsModule,
+        // ChartsModule,
+        NgChartsModule,
+        FormsModule,
+        FileUploadModule,
+        MatTableModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatTooltipModule,
+        MatButtonModule,
+        MatIconModule,
+        MatExpansionModule,
+        MatInputModule,
+        MatRadioModule,
+        MatDialogModule,
+        MatTabsModule,
+        MatSnackBarModule,
+        MatProgressSpinnerModule,
+        MatSelectModule,
+        ReactiveFormsModule,
+        MatFormFieldModule,
+        MatMenuModule,
+        MatCardModule,
+        FlexLayoutModule,
+        MatProgressBarModule,
+        MatToolbarModule,
+        MatBadgeModule,
+        MatTreeModule,
+        MatDividerModule,
+        MatCheckboxModule,
+        JwtModule.forRoot({
+            config: {
+                tokenGetter: getToken,
+                allowedDomains: ['api.aai.ebi.ac.uk'],
+                disallowedRoutes: ['https://api.aai.ebi.ac.uk/auth']
+            }
+        }),
+        // BsDropdownModule.forRoot(),
+        ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
+        NgxSmartModalModule.forRoot(),
+        // NgbModule,
+        NgcCookieConsentModule.forRoot(cookieConfig),
+        GraphQLModule,
+        MatListModule], providers: [ApiDataService, AggregationService, SlicePipe, UserService, ApiFiltersService, provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
