@@ -4,7 +4,8 @@ import { FilesSummaryComponent } from './files-summary.component';
 import {HeaderComponent} from '../shared/header/header.component';
 import {RouterTestingModule} from '@angular/router/testing';
 import {ChartsModule} from 'ng2-charts';
-import {HttpClientTestingModule} from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('FilesSummaryComponent', () => {
   let component: FilesSummaryComponent;
@@ -12,16 +13,14 @@ describe('FilesSummaryComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [
+    declarations: [
         FilesSummaryComponent,
         HeaderComponent
-      ],
-      imports: [
-        RouterTestingModule,
-        ChartsModule,
-        HttpClientTestingModule
-      ]
-    })
+    ],
+    imports: [RouterTestingModule,
+        ChartsModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
     .compileComponents();
   }));
 
