@@ -6,7 +6,7 @@ import {UserService} from '../../services/user.service';
 import {MatPaginator} from '@angular/material/paginator';
 import {MatSort} from '@angular/material/sort';
 import {MatTableDataSource} from '@angular/material/table';
-import {HttpClient, HttpEventType} from '@angular/common/http';
+import { HttpClient, HttpEventType } from '@angular/common/http';
 import {Observable, of as observableOf} from 'rxjs';
 
 @Component({
